Item dropping mechanic reworked

+++ Violet/include/c/prng.h
@@ -56,7 +56,7 @@ extern "C" {
      * @param rng a random number generator. If NULL is given, rnd16 is used
      * @return an index in [0, size-1]
     */
-    int choice(u32 *p, int p_size, u16 (*rng)());
+    size_t choice(u32 *p, size_t p_size, u16 (*rng)());
 
     /**
      * Chooses with probabilties derived from a scoring array using a softmax function with temperature 20 / 1024
@@ -67,7 +67,7 @@ extern "C" {
      * @param rng a random number generator. If NULL is given, rnd16 is used
      * @return an index in [0, size-1]
      **/
-    int softmax_choice(int *logits, int size, int min, int max, u16 (*rng)());
+    size_t softmax_choice(int *logits, size_t size, int min, int max, u16 (*rng)());
 
 
 #ifdef	__cplusplus
